src/rl_maze/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.2.0=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
asttokens=2.0.8=pypi_0
attrs=22.1.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.11.1=pypi_0
bleach=5.0.1=pypi_0
ca-certificates=2022.07.19=h06a4308_0
cachetools=5.2.0=pypi_0
certifi=2022.6.15=py38h06a4308_0
cffi=1.15.1=pypi_0
charset-normalizer=2.1.1=pypi_0
click=8.1.3=pypi_0
cloudpickle=2.1.0=pypi_0
cycler=0.11.0=pypi_0
cython=0.29.32=pypi_0
debugpy=1.6.3=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
dill=0.3.5.1=pypi_0
dm-control=1.0.7=pypi_0
dm-env=1.5=pypi_0
dm-tree=0.1.7=pypi_0
entrypoints=0.4=pypi_0
executing=1.0.0=pypi_0
fasteners=0.17.3=pypi_0
fastjsonschema=2.16.1=pypi_0
fonttools=4.37.1=pypi_0
glfw=2.5.5=pypi_0
google-auth=2.11.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
griddly=1.4.3=pypi_0
grpcio=1.48.1=pypi_0
gym=0.25.2=pypi_0
gym-minigrid=1.1.0=pypi_0
gym-notices=0.0.8=pypi_0
h5py=3.7.0=pypi_0
idna=3.3=pypi_0
imageio=2.21.2=pypi_0
importlib-metadata=4.12.0=pypi_0
importlib-resources=5.9.0=pypi_0
ipykernel=6.15.2=pypi_0
ipython=8.4.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.7.2=pypi_0
jedi=0.18.1=pypi_0
jinja2=3.1.2=pypi_0
jsonschema=4.15.0=pypi_0
jupyter-client=7.3.5=pypi_0
jupyter-contrib-core=0.4.0=pypi_0
jupyter-contrib-nbextensions=0.5.1=pypi_0
jupyter-core=4.11.1=pypi_0
jupyter-highlight-selected-word=0.2.0=pypi_0
jupyter-latex-envs=1.4.6=pypi_0
jupyter-nbextensions-configurator=0.5.0=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=1.1.1=pypi_0
kiwisolver=1.4.4=pypi_0
labmaze=1.0.5=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
lxml=4.9.1=pypi_0
markdown=3.4.1=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.5.3=pypi_0
matplotlib-inline=0.1.6=pypi_0
mistune=2.0.4=pypi_0
mjrl=1.0.0=pypi_0
mujoco=2.2.2=pypi_0
mujoco-py=2.1.2.14=pypi_0
nbclient=0.6.7=pypi_0
nbconvert=7.0.0=pypi_0
nbformat=5.4.0=pypi_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.5=pypi_0
networkx=2.8.6=pypi_0
notebook=6.4.12=pypi_0
numpy=1.23.2=pypi_0
oauthlib=3.2.0=pypi_0
odfpy=1.4.1=pypi_0
opencv-python=4.6.0.66=pypi_0
openssl=1.1.1q=h7f8727e_0
packaging=21.3=pypi_0
pandas=1.4.4=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.2.0=pypi_0
pip=22.1.2=py38h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
plotly=5.10.0=pypi_0
prometheus-client=0.14.1=pypi_0
prompt-toolkit=3.0.31=pypi_0
protobuf=3.19.4=pypi_0
psutil=5.9.1=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pybullet=3.2.5=pypi_0
pycparser=2.21=pypi_0
pygame=2.1.2=pypi_0
pygments=2.13.0=pypi_0
pyopengl=3.1.6=pypi_0
pyparsing=2.4.7=pypi_0
pyrsistent=0.18.1=pypi_0
python=3.8.13=h12debd9_0
python-dateutil=2.8.2=pypi_0
pytz=2022.2.1=pypi_0
pywavelets=1.3.0=pypi_0
pyyaml=6.0=pypi_0
pyzmq=23.2.1=pypi_0
qgrid=1.3.1=pypi_0
readline=8.1.2=h7f8727e_1
requests=2.28.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-image=0.16.2=pypi_0
scipy=1.9.1=pypi_0
send2trash=1.8.0=pypi_0
setuptools=63.4.1=py38h06a4308_0
six=1.16.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
sqlite=3.39.2=h5082296_0
stack-data=0.5.0=pypi_0
tabulate=0.8.10=pypi_0
tenacity=8.0.1=pypi_0
tensorboard=2.10.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.15.0=pypi_0
tifffile=2022.8.12=pypi_0
tinycss2=1.1.1=pypi_0
tk=8.6.12=h1ccaba5_0
torch=1.12.1=pypi_0
torchinfo=1.7.0=pypi_0
tornado=6.2=pypi_0
tqdm=4.64.1=pypi_0
traitlets=5.3.0=pypi_0
typing-extensions=4.3.0=pypi_0
urllib3=1.26.12=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.6.1=pypi_0
xz=5.2.5=h7f8727e_1
zipp=3.8.1=pypi_0
zlib=1.2.12=h7f8727e_2
